feat: Add table-specific state updaters to ensure field order
